Created pretty raw method in MafwPlaylistAdapter, to append multiple items in one query and added basic support for "contents-changed" signal in MafwPlaylistAdapter. Issuing a command to get items returns operation identifier, so it can be cancelled.
NowPlayingIndicator now manages its visibility on its own. This simplified some code, but there's no way to make sure it's hidden as long as we want. A method to inhibit the indicator could be added if necessary.

Applied singleton scheme to NowPlayingWindow and made it hide on close, so it's available for faster opening.

Fixed some segfaults, not sure which of them were there before and which were introduced from the changes in this commit, but I couldn't get it to crash so far.

Implemented "add to now playing" in the playlist browser using one query for multiple items.

Getting playlist items in NowPlayingWindow seemed somewhat broken and it's been reworked a bit. Still not perfect, though.

Probably some more small fixes.